remove TimerNavigationActions and RoutinesNavigationActions interfaces
