+++ features/timer/src/main/java/com/enricog/timer/navigation/TimerNavigationActions.kt
@@ -1,5 +1,19 @@
 package com.enricog.timer.navigation
 
-internal interface TimerNavigationActions {
-    suspend fun backToRoutines()
+import com.enricog.navigation.Navigator
+import com.enricog.navigation.routes.RoutinesRoute
+import com.enricog.navigation.routes.RoutinesRouteInput
+import com.enricog.navigation.routes.TimerRoute
+import javax.inject.Inject
+
+internal class TimerNavigationActions @Inject constructor(
+    private val navigator: Navigator
+) {
+
+    suspend fun backToRoutines() {
+        val routeNavigation = RoutinesRoute.navigate(RoutinesRouteInput) {
+            popUpTo(TimerRoute.name) { inclusive = true }
+        }
+        navigator.navigate(routeNavigation)
+    }
 }
